Fixed error in test programs.
These are not futures, just programmer error in the extern
declaration.  Thanks, Tom.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23448 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ test/functions/sungeun/myExternRecord.h
@@ -1,5 +1,5 @@
 struct _myExternRecord {
-  int x;
+  int64_t x;
 };
 
 typedef struct _myExternRecord myExternRecord;
